feat: add image input
