+++ src/App.tsx
@@ -4,7 +4,7 @@ import { Route, Routes } from 'react-router-dom';
 import About from './components/about/about';
 import Home from './components/homePage/HomePage';
 import Header from './components/header/header';
-import FormPage from './components/formPage/formPage';
+import FormPage from './components/formPage/form/form';
 
 function App() {
   return (
