+++ src/components/formPage/inputFile/inputFileValidation.ts
@@ -1,0 +1,6 @@
+export function inputFileValidation(image: File | undefined | null): boolean {
+  if (image) {
+    return image.type.includes('image');
+  }
+  return false;
+}
